scripts/python/python (github)/python_cloud/requirements.txt
azure-functions
azure-storage-blob
pyodbc
requests
urllib3